apps/api/requirements.txt
fastapi
uvicorn
sqlmodel
alembic
python-multipart
requests
pillow
numpy
